# Cache raw data points