--- a/Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
+++ b/Engine/DataFeeds/TextSubscriptionDataSourceReader.cs
@@ -155,7 +155,7 @@ namespace QuantConnect.Lean.Engine.DataFeeds
                     $"Type is: {cacheItem.Value.GetType()}");
             }
             // Find the first data point 10 days (just in case) before the desired date
-            // and subtract one item (just in case there wasn't a time gap and data.Time is after _date)
+            // and subtract one item (just in case there was a time gap and data.Time is after _date)
             var index = cache.FindIndex(data => data.Time.AddDays(10) > _date);
             index = index > 0 ? (index - 1) : 0;
             foreach (var data in cache.Skip(index))