<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- `TextSubscriptionDataSourceReader` will now cache raw price mode data
points for each requested source file, per data type.
   - Caching will be handled by the `MemoryCache` class. Max physical memory usage will be 70%. 
Cache entry should be evicted if it has not been accessed in 5 minutes.
   - Adding unit tests
- `GetLastKnownPrice()` will not `RoundDown` end time. This was causing
it to fail to fetch any data point in some cases.

### CLOUD performance testing
```
        public override void Initialize()
        {
            SetStartDate(2016, 01, 01);
            SetEndDate(2018, 01, 01);
            SetCash(10000);
            _symbol = AddEquity("SPY").Symbol;
        }
        public override void OnEndOfDay()
        {
            var dailyHistory = History(_symbol, 1, Resolution.Daily/Hour).First();
        }
```

**PR Resolution.Hour**
4.62 seconds at 42k data points per second. Processing total of 196,141 data points.
4.38 seconds at 45k data points per second. Processing total of 196,141 data points.
4.86 seconds at 40k data points per second. Processing total of 196,141 data points.
**PR Resolution.Daily**
5.13 seconds at 38k data points per second. Processing total of 196,141 data points.
4.27 seconds at 46k data points per second. Processing total of 196,141 data points.
4.02 seconds at 49k data points per second. Processing total of 196,141 data points.
**MASTER Resolution.Hour**
82.92 seconds at 2k data points per second. Processing total of 196,141 data points.
82.48 seconds at 2k data points per second. Processing total of 196,141 data points.
85.68 seconds at 2k data points per second. Processing total of 196,141 data points.
**MASTER Resolution.Daily**
16.73 seconds at 12k data points per second. Processing total of 196,141 data points.
17.46 seconds at 11k data points per second. Processing total of 196,141 data points.
15.71 seconds at 12k data points per second. Processing total of 196,141 data points.

**Result: ~20x improvement for `Resolution.Hour`. ~4x `Resolution.Daily`**

### Performance Analysis
**MASTER Resolution.Daily**
![imagen](https://user-images.githubusercontent.com/18473240/53649574-31498580-3c21-11e9-92a9-d3829e6f2939.png)

**MASTER Resolution.Hour**
![imagen](https://user-images.githubusercontent.com/18473240/53649799-c51b5180-3c21-11e9-973e-f843a8e231e0.png)

Both `Resolution.Daily` and `Resolution.Hour` data files are organized in 1 zip file per symbol (for the entire data history). **Each** history request has to open, read and parse this only data file. The objective of this PR is to cache and reuse this raw data.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/2820
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Daily and Hour Resolution History Requests were adding significant overhead to overall performance.
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests. Performance testing in the cloud. Live deployment. Local testing
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->